Move `toObject` and `fromObject` to a separate package
